Recipe #107 round 2: all 7 #61c instances -> 100. ObjModel_CopyJointTranslation (un-named pointer chain + ternary-phi temp: target's r0 merge = ternary expression in the consumer, #42 family), Obj_InsertIntoUpdateList (NEW lever: narrow-typed s16 local jumps the coloring queue like a temp -- retype to int with the *(s16*) load cast kept, decl/init order inert), dimlogfire_init (was actually #90 doubled-float-arg hoist, laundered). +708 matched_code, zero regressions, dol md5 intact

+++ src/main/object.c
@@ -597,7 +597,7 @@ void Obj_InsertIntoUpdateList(u8 *obj) {
         int *list = &lbl_803DCB7C;
         int prev = 0;
         int cur = list[1];
-        s16 linkOff = *(s16 *)((u8 *)list + 2);
+        int linkOff = *(s16 *)((u8 *)list + 2);
         while (cur != 0 && (s8)obj[0xae] < (s8)((u8 *)cur)[0xae]) {
             prev = cur;
             cur = *(int *)((u8 *)cur + linkOff);
